Move notebook cell logic in separate file
